Implement some memory pipe operations, and move async wrappers to the pipe mod
