+++ crates/wasi/src/preview2/mod.rs
@@ -37,7 +37,7 @@ pub use self::filesystem::{DirPerms, FilePerms};
 pub use self::poll::{HostPollable, TablePollableExt};
 pub use self::random::{thread_rng, Deterministic};
 pub use self::stream::{
-    AsyncReadStream, AsyncWriteStream, HostInputStream, HostOutputStream, StreamState,
+    HostInputStream, HostOutputStream, StreamState,
     TableStreamExt,
 };
 pub use self::table::{Table, TableError};
